+++ packages/backend/src/queryBuilder.test.ts
@@ -3,16 +3,23 @@ import {
     CustomDimensionType,
     ForbiddenError,
     isCustomBinDimension,
+    WeekDay,
 } from '@lightdash/common';
 import {
     assertValidDimensionRequiredAttribute,
     buildQuery,
     getCustomBinDimensionSql,
+    getCustomSqlDimensionSql,
     replaceUserAttributes,
+    sortDayOfWeekName,
+    sortMonthName,
 } from './queryBuilder';
 import {
     bigqueryClientMock,
     COMPILED_DIMENSION,
+    COMPILED_MONTH_NAME_DIMENSION,
+    COMPILED_WEEK_NAME_DIMENSION,
+    CUSTOM_SQL_DIMENSION,
     EXPLORE,
     EXPLORE_ALL_JOIN_TYPES_CHAIN,
     EXPLORE_BIGQUERY,
@@ -56,7 +63,9 @@ import {
     METRIC_QUERY_WITH_TABLE_CALCULATION_FILTER_SQL,
     METRIC_QUERY_WITH_TABLE_REFERENCE,
     METRIC_QUERY_WITH_TABLE_REFERENCE_SQL,
+    MONTH_NAME_SORT_SQL,
     warehouseClientMock,
+    WEEK_NAME_SORT_SQL,
 } from './queryBuilder.mock';
 
 describe('Query builder', () => {
@@ -525,6 +534,18 @@ describe('with custom dimensions', () => {
         ).toStrictEqual(undefined);
     });
 
+    it('getCustomSqlDimensionSql with custom sql dimension', () => {
+        expect(
+            getCustomSqlDimensionSql({
+                warehouseClient: bigqueryClientMock,
+                customDimensions: [CUSTOM_SQL_DIMENSION],
+            }),
+        ).toStrictEqual({
+            selects: ['  ("table1".dim1 < 18) AS `is_adult`'],
+            tables: ['table1'],
+        });
+    });
+
     it('getCustomDimensionSql with custom dimension', () => {
         expect(
             getCustomBinDimensionSql({
@@ -532,7 +553,7 @@ describe('with custom dimensions', () => {
 
                 explore: EXPLORE,
                 customDimensions:
-                    METRIC_QUERY_WITH_CUSTOM_DIMENSION.customDimensions?.filter(
+                    METRIC_QUERY_WITH_CUSTOM_DIMENSION.compiledCustomDimensions?.filter(
                         isCustomBinDimension,
                     ),
                 userAttributes: {},
@@ -643,7 +664,7 @@ LIMIT 10`);
                 explore: EXPLORE,
                 compiledMetricQuery: {
                     ...METRIC_QUERY_WITH_CUSTOM_DIMENSION,
-                    customDimensions: [
+                    compiledCustomDimensions: [
                         {
                             id: 'age_range',
                             name: 'Age range',
@@ -738,7 +759,7 @@ LIMIT 10`);
                 warehouseClient: bigqueryClientMock,
                 explore: EXPLORE,
                 customDimensions:
-                    METRIC_QUERY_WITH_CUSTOM_DIMENSION.customDimensions?.filter(
+                    METRIC_QUERY_WITH_CUSTOM_DIMENSION.compiledCustomDimensions?.filter(
                         isCustomBinDimension,
                     ),
                 userAttributes: {},
@@ -828,7 +849,7 @@ LIMIT 10`);
                 explore: EXPLORE,
                 compiledMetricQuery: {
                     ...METRIC_QUERY_WITH_CUSTOM_DIMENSION,
-                    customDimensions: [
+                    compiledCustomDimensions: [
                         {
                             id: 'age_range',
                             name: 'Age range',
@@ -854,5 +875,52 @@ FROM "db"."schema"."table1" AS "table1"
 GROUP BY 1,2
 ORDER BY "table1_metric1" DESC
 LIMIT 10`);
+    });
+});
+
+const ignoreIndentation = (sql: string) => sql.replace(/\s+/g, ' ');
+describe('Time frame sorting', () => {
+    it('sortMonthName SQL', () => {
+        expect(
+            ignoreIndentation(sortMonthName(COMPILED_MONTH_NAME_DIMENSION)),
+        ).toStrictEqual(ignoreIndentation(MONTH_NAME_SORT_SQL));
+    });
+    it('sortDayOfWeekName SQL for undefined startOfWeek', () => {
+        expect(
+            ignoreIndentation(
+                sortDayOfWeekName(COMPILED_WEEK_NAME_DIMENSION, undefined),
+            ),
+        ).toStrictEqual(ignoreIndentation(WEEK_NAME_SORT_SQL));
+    });
+    it('sortDayOfWeekName SQL for Sunday startOfWeek', () => {
+        expect(
+            ignoreIndentation(
+                sortDayOfWeekName(COMPILED_WEEK_NAME_DIMENSION, WeekDay.SUNDAY),
+            ),
+        ).toStrictEqual(ignoreIndentation(WEEK_NAME_SORT_SQL)); // same as undefined
+    });
+
+    it('sortDayOfWeekName SQL for Wednesday startOfWeek', () => {
+        expect(
+            ignoreIndentation(
+                sortDayOfWeekName(
+                    COMPILED_WEEK_NAME_DIMENSION,
+                    WeekDay.WEDNESDAY,
+                ),
+            ),
+        ).toStrictEqual(
+            ignoreIndentation(`(
+            CASE
+                WHEN "table1".dim1 = 'Sunday' THEN 5
+                WHEN "table1".dim1 = 'Monday' THEN 6
+                WHEN "table1".dim1 = 'Tuesday' THEN 7
+                WHEN "table1".dim1 = 'Wednesday' THEN 1
+                WHEN "table1".dim1 = 'Thursday' THEN 2
+                WHEN "table1".dim1 = 'Friday' THEN 3
+                WHEN "table1".dim1 = 'Saturday' THEN 4
+                ELSE 0
+            END
+        )`),
+        );
     });
 });